Fixed tests after refactorings
